Revert "trivial type cast from pid_t to int in printf() alike functions."
This reverts commit 08b2e8f6af24f37bcec06e552de7e1c81e347c68.

+++ src/common.h
@@ -13,9 +13,9 @@ void common_print_backtrace(void);
 
 #ifdef PRINTF_DEBUG
   #include <stdio.h>
-  #define	DPRINTF(level, fmt, args...)	{ fprintf(stderr, "%d->%s: " fmt, (int) getpid(), __FUNCTION__, ## args); fflush(stderr); }
+  #define	DPRINTF(level, fmt, args...)	{ fprintf(stderr, "%d->%s: " fmt, getpid(), __FUNCTION__, ## args); fflush(stderr); }
 #else
-  #define	DPRINTF(level, fmt, args...)	common_debug_print(level, "%d->%s: " fmt, (int) getpid(), __FUNCTION__, ## args)
+  #define	DPRINTF(level, fmt, args...)	common_debug_print(level, "%d->%s: " fmt, getpid(), __FUNCTION__, ## args)
 #endif
 
 #ifndef HAVE_STRNDUP
